# Remove unpredicted branch from kmerge::sift_down